# docs/openapi/swagger.yaml
components:
  schemas:
...
    github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-internal_api_handler_plugin_detailResponse:
      properties:
        data:
          $ref: '#/components/schemas/internal_api_handler_plugin.detailResponse'
      type: object
...
    github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error:
      properties:
        error:
          $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.ErrorBody'
      type: object
    github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.ErrorBody:
      properties:
        code:
          type: string
        details:
          additionalProperties: {}
          type: object
        hint:
          type: string
        message:
          type: string
      type: object
...
    github_com_Mininglamp-OSS_octo-marketplace_internal_model.PluginDisplayStatus:
      description: 'DisplayStatus is the single DERIVED status a client shows: draft
        /

        pending_review / published / rejected / delisted. It folds the listing axis

        together with the review axis so clients do not have to fetch the review list

        and join. Distinct from `status`, which is the row''s soft-active flag.'
      enum:
      - draft
      - pending_review
      - published
      - rejected
      - delisted
      type: string
      x-enum-comments:
        PluginDisplayStatusDelisted: 已下架
        PluginDisplayStatusDraft: 草稿
        PluginDisplayStatusPendingReview: 审核中
        PluginDisplayStatusPublished: 已发布
        PluginDisplayStatusRejected: 驳回
      x-enum-varnames:
      - PluginDisplayStatusDraft
      - PluginDisplayStatusPendingReview
      - PluginDisplayStatusPublished
      - PluginDisplayStatusRejected
      - PluginDisplayStatusDelisted
    github_com_Mininglamp-OSS_octo-marketplace_internal_model.PluginListingState:
      description: 'PluginListingState is the plugin''s CURRENT listing state, which
        can have

        moved on since this request was decided — an approved plugin a Space admin

        later delisted reads `delisted`. The review queue renders the row status from

        it and uses it to stop offering 下架 on something already down.'
      enum:
      - draft
      - published
      - delisted
      type: string
      x-enum-varnames:
      - PluginListingStateDraft
      - PluginListingStatePublished
      - PluginListingStateDelisted
    github_com_Mininglamp-OSS_octo-marketplace_internal_model.PluginType:
      enum:
      - expert
      - expert_team
      - skill
      - connector
      type: string
      x-enum-varnames:
      - PluginTypeExpert
      - PluginTypeExpertTeam
      - PluginTypeSkill
      - PluginTypeConnector
    github_com_Mininglamp-OSS_octo-marketplace_internal_model.PluginVisibility:
      enum:
      - public
      - space
      - private
      - system
      type: string
      x-enum-varnames:
      - PluginVisibilityPublic
      - PluginVisibilitySpace
      - PluginVisibilityPrivate
      - PluginVisibilitySystem
...
    internal_api_handler_plugin.detailResponse:
      properties:
        plugin:
          $ref: '#/components/schemas/internal_api_handler_plugin.pluginResponse'
        relation_result:
          $ref: '#/components/schemas/internal_api_handler_plugin.relationResultResponse'
        relations:
          items:
            $ref: '#/components/schemas/internal_api_handler_plugin.relationResponse'
          type: array
          uniqueItems: false
      type: object
...
    internal_api_handler_plugin.pluginResponse:
      properties:
        category_id:
          type: string
        created_at:
          type: string
        created_by_bot_id:
          type: string
        created_by_bot_name:
          type: string
        created_by_type:
          type: string
        creator_name:
          type: string
        current_version:
          type: string
        current_version_id:
          type: string
        display_status:
          $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_model.PluginDisplayStatus'
        download_count:
          type: integer
        icon:
          description: 'Icon is the stored write-canonical value clients must echo
            on updates;

            IconURL is the resolved display URL (presigned when Icon is an object
            key).'
          type: string
        icon_url:
          type: string
        install_count:
          type: integer
        is_embedded:
          type: boolean
        listing_state:
          $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_model.PluginListingState'
        manifest_hash:
          type: string
        manifest_json:
          type: object
        owner_id:
          type: string
        plugin_hash:
          type: string
        plugin_id:
          type: string
        plugin_json:
          type: object
        plugin_name:
          type: string
        plugin_type:
          $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_model.PluginType'
        publisher:
          type: string
        review_id:
          type: string
        space_id:
          type: string
        status:
          type: integer
        tags:
          items:
            type: string
          type: array
          uniqueItems: false
        tool_count:
          type: integer
        updated_at:
          type: string
        view_count:
          type: integer
        visibility:
          $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_model.PluginVisibility'
      type: object
...
    internal_api_handler_plugin.relationResponse:
      properties:
        data:
          type: object
        relation_id:
          type: string
        relation_type:
          type: string
        sort_order:
          type: integer
        source_plugin_id:
          type: string
        target_plugin_id:
          type: string
      type: object
    internal_api_handler_plugin.relationResultResponse:
      description: 'RelationResult reports upsert relation target-state synchronization:

        empty relation_id created, known relation_id with changes updated, live

        relations omitted from the submission deleted. Absent on reads.'
      properties:
        created:
          items:
            type: string
          type: array
          uniqueItems: false
        deleted:
          items:
            type: string
          type: array
          uniqueItems: false
        updated:
          items:
            type: string
          type: array
          uniqueItems: false
      type: object
...
  securitySchemes:
    AdminToken:
      in: header
      name: X-Admin-Token
      type: apiKey
    Bearer:
      scheme: bearer
      type: http
info:
  contact:
    name: OCTO API Team
    url: https://github.com/Mininglamp-OSS/octo-marketplace
  description: 'Skill and MCP marketplace API for OCTO.


    A 409 CONFLICT carries `error.details.conflict_reason`, which is what a

    client should branch on — the status code alone does not distinguish these:

    `review_pending` (a request is already open on this plugin; cancel it first),

    `listed_requires_review` (the plugin is listed to the organization, so the

    change must go through review or an admin must delist it first),

    `label_taken` (the submitted version label is already published; pick another

    and resubmit), `already_published` / `not_published` (the listing transition

    does not apply in the plugin''s current state), `deadlock` (transient lock

    contention; also sets `retryable: true`, so retry the request unchanged), and

    `state` (a generic raced state change; refresh and retry).'
  title: Octo Marketplace API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /plugins/review_requests/{review_id}/approve:
    post:
      description: Requires the Space owner/admin role. Applies the frozen documents
        and relation graph to the plugin, records a release version carrying the applicant's
        label, and for a first listing flips visibility from private to space. Losing
        a concurrent decision race returns CONFLICT.
      operationId: plugin.review_request.approve
      parameters:
      - description: Review request ID
        in: path
        name: review_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-internal_api_handler_plugin_detailResponse'
          description: OK
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: VALIDATION_ERROR
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: AUTH_REQUIRED
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: 'FORBIDDEN: Space owner/admin role required'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: NOT_FOUND
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: 'CONFLICT: already decided'
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: RATE_LIMITED
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: INTERNAL_ERROR
      security:
      - Bearer: []
      summary: Approve a plugin review request
      tags:
      - plugin
...
servers:
- url: /v1
...